requirements.txt
amulet-leveldb>=1.0.2
rapidnbt>=1.1.0